[Console] Add ContainerAwareCommand
